# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Buccaneer Wallet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /owner-auth/bootstrap/device:
    post:
      operationId: bootstrapOwnerDevice
      tags: [owner-auth]
      summary: Register or relabel an owner device
      security:
        - ownerBootstrapHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapOwnerDeviceRequest"
      responses:
        "200":
          description: Owner device updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerDevice"
        "201":
          description: Owner device created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerDevice"
  /owner-auth/register-initial-device:
    post:
      operationId: registerInitialOwnerDevice
      tags: [owner-auth]
      summary: Register the first trusted owner device when none exists yet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapOwnerDeviceRequest"
      responses:
        "201":
          description: First owner device created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerDevice"
  /owner-auth/challenge:
    post:
      operationId: createOwnerAuthChallenge
      tags: [owner-auth]
      summary: Create an owner auth challenge for a registered device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerAuthChallengeRequest"
      responses:
        "200":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerAuthChallengeResponse"
  /owner-auth/verify:
    post:
      operationId: verifyOwnerAuthChallenge
      tags: [owner-auth]
      summary: Verify a signed owner auth challenge and issue a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerAuthVerifyRequest"
      responses:
        "200":
          description: Owner session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerSessionResponse"
  /owner-auth/logout:
    post:
      operationId: logoutOwnerSession
      tags: [owner-auth]
      summary: Revoke the current owner session
      security:
        - ownerSessionHeader: []
      responses:
        "200":
          description: Owner session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /owner-auth/status:
    get:
      operationId: getOwnerSessionStatus
      tags: [owner-auth]
      summary: Get the current owner session status
      security:
        - ownerSessionHeader: []
      responses:
        "200":
          description: Active owner session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerSessionStatus"
...
components:
  securitySchemes:
    ownerSessionHeader:
      type: apiKey
      in: header
      name: X-Owner-Session
    ownerBootstrapHeader:
      type: apiKey
      in: header
      name: X-Wallet-Owner
  schemas:
...
    BootstrapOwnerDeviceRequest:
      type: object
      properties:
        publicKey:
          type: string
        label:
          type: string
      required: [publicKey, label]
    OwnerDevice:
      type: object
      properties:
        id:
          type: integer
        publicKey:
          type: string
        label:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, publicKey, label, createdAt]
    OwnerAuthChallengeRequest:
      type: object
      properties:
        publicKey:
          type: string
      required: [publicKey]
    OwnerAuthChallengeResponse:
      type: object
      properties:
        challengeId:
          type: integer
        nonce:
          type: string
        expiresAt:
          type: string
          format: date-time
        deviceLabel:
          type: string
        serverTime:
          type: string
          format: date-time
      required: [challengeId, nonce, expiresAt, deviceLabel, serverTime]
    OwnerAuthVerifyRequest:
      type: object
      properties:
        challengeId:
          type: integer
        publicKey:
          type: string
        signature:
          type: string
      required: [challengeId, publicKey, signature]
    OwnerSessionResponse:
      type: object
      properties:
        sessionToken:
          type: string
        expiresAt:
          type: string
          format: date-time
        device:
          type: object
          properties:
            id:
              type: integer
            label:
              type: string
            publicKey:
              type: string
          required: [id, label, publicKey]
      required: [sessionToken, expiresAt, device]
    OwnerSessionStatus:
      type: object
      properties:
        ok:
          type: boolean
        session:
          type: object
          properties:
            sessionId:
              type: integer
            expiresAt:
              type: string
              format: date-time
            deviceId:
              type: integer
            label:
              type: string
            publicKey:
              type: string
          required: [sessionId, expiresAt, deviceId, label, publicKey]
      required: [ok, session]
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...